Fix integrations message example (#14111)
