+++ app/integrations/client/views/integrationsIncoming.html
@@ -116,7 +116,7 @@
 							</div>
 						</div>
 						<div class="input-line message-example">
-							{{#nrr nrrargs 'message' example}}{{/nrr}}
+							{{#nrr nrrargs 'message' msg=exampleMsg settings=exampleSettings u=exampleUser}}{{/nrr}}
 						</div>
 					</div>
 				</div>
