+++ app/integrations/client/views/integrationsIncoming.js
@@ -1,6 +1,5 @@
 import { Meteor } from 'meteor/meteor';
 import { ReactiveVar } from 'meteor/reactive-var';
-import { Random } from 'meteor/random';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { Template } from 'meteor/templating';
 import { TAPi18n } from 'meteor/tap:i18n';
@@ -8,6 +7,7 @@ import { hasAtLeastOnePermission, hasAllPermission } from '../../../authorizatio
 import { modal } from '../../../ui-utils';
 import { t, handleError } from '../../../utils';
 import { ChatIntegrations } from '../collections';
+import { exampleMsg, exampleSettings, exampleUser } from './messageExample';
 import hljs from 'highlight.js';
 import toastr from 'toastr';
 
@@ -18,6 +18,9 @@ Template.integrationsIncoming.onCreated(function _incomingIntegrationsOnCreated(
 });
 
 Template.integrationsIncoming.helpers({
+	exampleMsg,
+	exampleUser,
+	exampleSettings,
 	hasPermission() {
 		return hasAtLeastOnePermission(['manage-integrations', 'manage-own-integrations']);
 	},
@@ -44,34 +47,6 @@ Template.integrationsIncoming.helpers({
 
 		return Template.instance().record.curValue;
 	},
-
-	example() {
-		const record = Template.instance().record.get();
-		return {
-			_id: Random.id(),
-			alias: record.alias,
-			emoji: record.emoji,
-			avatar: record.avatar,
-			msg: 'Example message',
-			bot: {
-				i: Random.id(),
-			},
-			groupable: false,
-			attachments: [{
-				title: 'Rocket.Chat',
-				title_link: 'https://rocket.chat',
-				text: 'Rocket.Chat, the best open source chat',
-				image_url: '/images/integration-attachment-example.png',
-				color: '#764FA5',
-			}],
-			ts: new Date(),
-			u: {
-				_id: Random.id(),
-				username: record.username,
-			},
-		};
-	},
-
 	exampleJson() {
 		const record = Template.instance().record.get();
 		const data = {
