+++ app/integrations/client/views/integrationsOutgoing.html
@@ -171,7 +171,7 @@
 									</div>
 								</div>
 								<div class="input-line message-example">
-									{{#nrr nrrargs 'message' example}}{{/nrr}}
+									{{#nrr nrrargs 'message' msg=exampleMsg settings=exampleSettings u=exampleUser}}{{/nrr}}
 								</div>
 							{{/if}}
 						</div>
